refactor: TableGroup 검증 로직을 도메인 안으로 이동
