+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -4,7 +4,6 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
@@ -12,6 +11,7 @@ import kitchenpos.dao.TableGroupDao;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.Tables;
 import kitchenpos.dto.OrderTableResponse;
 import kitchenpos.dto.TableGroupRequest;
 import kitchenpos.dto.TableGroupResponse;
@@ -33,32 +33,14 @@ public class TableGroupService {
 
     @Transactional
     public TableGroupResponse create(final TableGroupRequest tableGroupRequest) {
-        if (tableGroupRequest.getOrderTables().size() < 2) {
-            throw new IllegalArgumentException("등록되는 테이블 수가 2 이상이어야 한다.");
-        }
+        final Tables orderTables = new Tables(getOrderTables(tableGroupRequest));
 
-        final List<OrderTable> orderTables = getOrderTables(tableGroupRequest);
-        validateNoGroupedTable(orderTables);
-        validateTableIsEmpty(orderTables);
-
-        final TableGroup savedTableGroup = tableGroupDao.save(new TableGroup(null, LocalDateTime.now(), orderTables));
-
+        final TableGroup savedTableGroup = tableGroupDao.save(
+                new TableGroup(null, LocalDateTime.now(), orderTables.getOrderTables()));
         updateToFull(savedTableGroup);
         addTableGroupId(savedTableGroup);
 
         return toTableGroupResponse(savedTableGroup);
-    }
-
-    private void validateTableIsEmpty(List<OrderTable> orderTables) {
-        if (orderTables.stream().anyMatch(orderTable -> !orderTable.isEmpty())) {
-            throw new IllegalArgumentException("등록되는 모든 테이블들은 비어있어야 한다.");
-        }
-    }
-
-    private void validateNoGroupedTable(List<OrderTable> orderTables) {
-        if (orderTables.stream().anyMatch(orderTable -> Objects.nonNull(orderTable.getTableGroupId()))) {
-            throw new IllegalArgumentException("등록되는 모든 테이블들은 기존 단체 지정이 없어야 한다.");
-        }
     }
 
     private List<OrderTable> getOrderTables(TableGroupRequest tableGroupRequest) {
