+++ src/main/java/kitchenpos/dao/jdbctemplate/JdbcTemplateTableGroupDao.java
@@ -3,7 +3,6 @@ package kitchenpos.dao.jdbctemplate;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import javax.sql.DataSource;
@@ -60,7 +59,6 @@ public class JdbcTemplateTableGroupDao {
 
     private TableGroup toEntity(final ResultSet resultSet) throws SQLException {
         return new TableGroup(resultSet.getLong(KEY_COLUMN_NAME),
-                resultSet.getObject("created_date", LocalDateTime.class),
-                new ArrayList<>());
+                resultSet.getObject("created_date", LocalDateTime.class));
     }
 }
